+++ app/src/main/java/com/capstone/designpatterntutorial/model/mainscreen/Pattern.kt
@@ -4,9 +4,10 @@ import java.io.Serializable
 
 data class Pattern(
     val id: Int,
+    val categoryId: Int,
     val name: String,
     val summary: String,
     val url: String,
-    val type: String,
-    val isFavorite: Boolean = false
+    val imageName: String,
+    var isFavorite: Boolean = false
 ) : Serializable